requirements.txt
Flask==1.1.2
WTForms==2.3.3
SQLAlchemy==1.4.5
PyJWT==2.1.0
Werkzeug==1.0.1
alembic==1.6.1
eventlet==0.31.0
python-dotenv==0.17.1